refactor: move analytics to component
