+++ apps/web/src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { Analytics } from "@vercel/analytics/react";
+import { Analytics as VercelAnalytics } from "@vercel/analytics/react";
 import "../styles/globals.css";
 
 import { Inter } from "next/font/google";
@@ -6,28 +6,13 @@ const inter = Inter({ subsets: ["latin"] });
 
 import type { AppProps } from "next/app";
 import { Hydrate, QueryClient, QueryClientProvider } from "react-query";
+import { Analytics } from "../components/Analytics";
 import { Navbar } from "../components/Navbar";
 import WagmiProvider from "../components/Web3Provider";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
-import * as analytics from "../lib/analytics";
 
 const queryClient = new QueryClient({});
 
 function MyApp({ Component, pageProps }: AppProps) {
-  const router = useRouter();
-
-  useEffect(() => {
-    const handler = (url: string) => {
-      analytics.page(url);
-    };
-
-    router.events.on("routeChangeComplete", handler);
-    return () => {
-      router.events.off("routeChangeComplete", handler);
-    };
-  }, [router.events]);
-
   return (
     <QueryClientProvider client={queryClient}>
       <Hydrate state={pageProps.dehydratedState}>
@@ -39,6 +24,7 @@ function MyApp({ Component, pageProps }: AppProps) {
             <Component {...pageProps} />
           </main>
           <Analytics />
+          <VercelAnalytics />
         </WagmiProvider>
       </Hydrate>
     </QueryClientProvider>
